Kanren.LogicalBase

Description

Logical representations for some base types along with their (orphan) Logical instances.

Primitive types

There are Logical instances for Bool, Char, Int, and Void. They don't require a separate logic representation.

Prisms for Bool

_False :: Prism' Bool ()

_True :: Prism' Bool ()

_False' :: ExhaustivePrism Bool (false, true) (false', true) () false false'

_True' :: ExhaustivePrism Bool (false, true) (false, true') () true true'

Tuples

For tuples, the logical representation is a tuple too, so they don't need a separate logic type. Currently, there's a Logical instance for 2-tuples only.

Lists

data LogicList a

Constructors

LogicNil

LogicCons (Term a) (Term [a])

Prisms for lists

_LogicNil :: forall a. Prism' (LogicList a) ()

_LogicCons :: forall a a. Prism (LogicList a) (LogicList a) (Term a, Term [a]) (Term a, Term [a])

_LogicNil' :: ExhaustivePrism (LogicList a) (logicNil, logicCons) (logicNil', logicCons) () logicNil logicNil'

_LogicCons' :: ExhaustivePrism (LogicList a) (logicNil, logicCons) (logicNil, logicCons') (Term a, Term [a]) logicCons logicCons'

Maybe

data LogicMaybe (a :: Type)

Constructors

LogicNothing

LogicJust (Term a)

Prisms for Maybe

_LogicNothing :: forall a. Prism' (LogicMaybe a) ()

_LogicJust :: forall a a. Prism (LogicMaybe a) (LogicMaybe a) (Term a) (Term a)

_LogicNothing' :: ExhaustivePrism (LogicMaybe a) (logicNothing, logicJust) (logicNothing', logicJust) () logicNothing logicNothing'

_LogicJust' :: ExhaustivePrism (LogicMaybe a) (logicNothing, logicJust) (logicNothing, logicJust') (Term a) logicJust logicJust'

Either

data LogicEither (a :: Type) (b :: Type)

Constructors

LogicLeft (Term a)

LogicRight (Term b)

Prisms for Either

_LogicLeft :: forall a b a. Prism (LogicEither a b) (LogicEither a b) (Term a) (Term a)

_LogicRight :: forall a b b. Prism (LogicEither a b) (LogicEither a b) (Term b) (Term b)

_LogicLeft' :: ExhaustivePrism (LogicEither a b) (logicLeft, logicRight) (logicLeft', logicRight) (Term a) logicLeft logicLeft'

_LogicRight' :: ExhaustivePrism (LogicEither a b) (logicLeft, logicRight) (logicLeft, logicRight') (Term b) logicRight logicRight'

Orphan instances

Logical Void

Instance details

Associated Types

type Logic Void = (r :: Type)

Methods

unify :: Logic Void -> Logic Void -> State -> Maybe State

walk :: State -> Logic Void -> Logic Void

occursCheck :: VarId b -> Logic Void -> State -> Bool

inject :: Void -> Logic Void

extract :: Logic Void -> Maybe Void

Logical Bool

Instance details

Associated Types

type Logic Bool = (r :: Type)

Methods

unify :: Logic Bool -> Logic Bool -> State -> Maybe State

walk :: State -> Logic Bool -> Logic Bool

occursCheck :: VarId b -> Logic Bool -> State -> Bool

inject :: Bool -> Logic Bool

extract :: Logic Bool -> Maybe Bool

Logical Char

Instance details

Associated Types

type Logic Char = (r :: Type)

Methods

unify :: Logic Char -> Logic Char -> State -> Maybe State

walk :: State -> Logic Char -> Logic Char

occursCheck :: VarId b -> Logic Char -> State -> Bool

inject :: Char -> Logic Char

extract :: Logic Char -> Maybe Char

Logical Int

Instance details

Associated Types

type Logic Int = (r :: Type)

Methods

unify :: Logic Int -> Logic Int -> State -> Maybe State

walk :: State -> Logic Int -> Logic Int

occursCheck :: VarId b -> Logic Int -> State -> Bool

inject :: Int -> Logic Int

extract :: Logic Int -> Maybe Int

Logical a => Logical (Maybe a)

Instance details

Associated Types

type Logic (Maybe a) = (r :: Type)

Methods

unify :: Logic (Maybe a) -> Logic (Maybe a) -> State -> Maybe State

walk :: State -> Logic (Maybe a) -> Logic (Maybe a)

occursCheck :: VarId b -> Logic (Maybe a) -> State -> Bool

inject :: Maybe a -> Logic (Maybe a)

extract :: Logic (Maybe a) -> Maybe (Maybe a)

Logical a => Logical [a]

Instance details

Associated Types

type Logic [a] = (r :: Type)

Methods

unify :: Logic [a] -> Logic [a] -> State -> Maybe State

walk :: State -> Logic [a] -> Logic [a]

occursCheck :: VarId b -> Logic [a] -> State -> Bool

inject :: [a] -> Logic [a]

extract :: Logic [a] -> Maybe [a]

(Logical a, Logical b) => Logical (Either a b)

Instance details

Associated Types

type Logic (Either a b) = (r :: Type)

Methods

unify :: Logic (Either a b) -> Logic (Either a b) -> State -> Maybe State

walk :: State -> Logic (Either a b) -> Logic (Either a b)

occursCheck :: VarId b0 -> Logic (Either a b) -> State -> Bool

inject :: Either a b -> Logic (Either a b)

extract :: Logic (Either a b) -> Maybe (Either a b)

(Logical a, Logical b) => Logical (a, b)

Instance details

Associated Types

type Logic (a, b) = (r :: Type)

Methods

unify :: Logic (a, b) -> Logic (a, b) -> State -> Maybe State

walk :: State -> Logic (a, b) -> Logic (a, b)

occursCheck :: VarId b0 -> Logic (a, b) -> State -> Bool

inject :: (a, b) -> Logic (a, b)

extract :: Logic (a, b) -> Maybe (a, b)