Kanren.TH

Description

Automatic generation of logic types.

Documentation

makeLogic :: Name -> Q [Dec]

Generate a logic representation and a Logical instance for the given type.

This function can be used with data- and newtypes. In the generated logic type,

the type name and all constructors will be prefixed with “Logic”;

each field will be wrapped in a Term;

each field accessor foo will be prefixed with “logic": logicFoo.

The generated Logical instance relies on the GenericLogical module. As such, the given type should have a Generic instance, and it will be derived for the logic type. For each type variable a in the type definition, there will be a Logical a constraint in the instance.

Take this type definition as an example:

data Tree a
  = Leaf a
  | Node (Tree a) (Tree a)
  deriving (Eq, Show, Generic)

makeLogic ''Tree will produce:

data LogicTree a
  = LogicLeaf (Term a)
  | LogicNode (Term (Tree a)) (Term (Tree a))
  deriving (Generic)

instance Logical a => Logical (Tree a) where
  type Logic (Tree a) = LogicTree a
  unify = genericUnify
  subst = genericSubst
  inject = genericInject
  extract = genericExtract

makeLogic requires the following extensions (besides TemplateHaskell):

ConstraintKinds;

DeriveGeneric;

TypeFamilies.